Change score type from Double to Int for consistency in player score representation

+++ gk-common/src/commonMain/kotlin/galkon/common/Domain.kt
@@ -171,7 +171,7 @@ data class PlayerScore(
     val player: PlayerInfo,
     val planetsOwned: Int,
     val totalShips: ShipCount,
-    val score: Double,
+    val score: Int,
 )
 
 /** Player info for status/lobby display. */
