+++ gk-common/src/commonMain/kotlin/galkon/common/Serialization.kt
@@ -159,7 +159,7 @@ data class PlayerScoreDto(
     val playerName: String,
     val planetsOwned: Int,
     val totalShips: Int,
-    val score: Double,
+    val score: Int,
 )
 
 @Serializable
